@@ -4,58 +4,70 @@
   "license": "MIT",
   "angular-cli": {},
   "scripts": {
-    "ng": "ng",
     "start": "ng serve --hmr -e=dev-hmr",
-    "serve": "ng serve",
+    "build-aot-and-deploy": "npm run build-aot && npm run deploy",
     "test": "ng test",
-    "lint": "ng lint",
     "e2e": "ng e2e",
+    "lint": "ng lint",
+    "ng": "ng",
+    "serve": "ng serve",
     "build": "ng build --prod",
     "build-aot": "ng build --prod --aot",
-    "deploy": "npm run build-aot && gh-pages -d dist -b master"
+    "deploy": "gh-pages -d dist -b master"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.1.3",
-    "@angular/common": "^4.1.3",
-    "@angular/compiler": "^4.1.3",
-    "@angular/core": "^4.1.3",
-    "@angular/forms": "^4.1.3",
-    "@angular/http": "^4.1.3",
-    "@angular/material": "^2.0.0-beta.6",
-    "@angular/platform-browser": "^4.1.3",
-    "@angular/platform-browser-dynamic": "^4.1.3",
-    "@angular/router": "^4.1.3",
+    "@angular/animations": "^4.3.2",
+    "@angular/common": "^4.3.2",
+    "@angular/compiler": "^4.3.2",
+    "@angular/core": "^4.3.2",
+    "@angular/forms": "^4.3.2",
+    "@angular/http": "^4.3.2",
+    "@angular/material": "^2.0.0-beta.8",
+    "@angular/platform-browser": "^4.3.2",
+    "@angular/platform-browser-dynamic": "^4.3.2",
+    "@angular/router": "^4.3.2",
+    "@ngrx/core": "^1.2.0",
+    "@ngrx/db": "^2.0.2",
+    "@ngrx/effects": "^4.0.1",
+    "@ngrx/router-store": "^4.0.0",
+    "@ngrx/store": "^4.0.0",
+    "@ngrx/store-devtools": "^4.0.0",
     "angular2-jwt": "^0.2.3",
-    "auth0-js": "^8.7.0",
+    "auth0-js": "^8.8.0",
     "core-js": "^2.4.1",
     "global": "^4.3.2",
     "hammerjs": "^2.0.8",
-    "ng2-md-datatable": "^1.3.15",
-    "rxjs": "^5.3.0",
-    "zone.js": "^0.8.12"
+    "ng2-md-datatable": "^1.4.0",
+    "reselect": "^3.0.1",
+    "rxjs": "^5.4.2",
+    "ts-helpers": "^1.1.2",
+    "zone.js": "^0.8.16"
   },
   "devDependencies": {
-    "@angular/cli": "^1.1.1",
-    "@angular/compiler-cli": "^4.1.3",
-    "@angular/language-service": "^4.2.0-rc.2",
-    "@angularclass/hmr": "^1.2.2",
-    "@types/jasmine": "^2.5.48",
-    "@types/node": "^7.0.29",
-    "codelyzer": "^3.0.1",
-    "eslint": "^3.19.0",
+    "@angular/cdk": "2.0.0-beta.8",
+    "@angular/cli": "^1.2.6",
+    "@angular/compiler-cli": "^4.3.2",
+    "@angular/language-service": "^4.3.2",
+    "@angularclass/hmr": "^2.1.3",
+    "@types/jasmine": "^2.5.53",
+    "@types/node": "^8.0.17",
+    "codelyzer": "^3.1.2",
+    "common-tags": "^1.4.0",
+    "eslint": "^4.3.0",
     "gh-pages": "^1.0.0",
-    "jasmine-core": "^2.6.3",
-    "jasmine-spec-reporter": "^4.1.0",
-    "karma": "^1.6.0",
-    "karma-chrome-launcher": "^2.1.1",
+    "jasmine-core": "^2.7.0",
+    "jasmine-spec-reporter": "^4.1.1",
+    "karma": "^1.7.0",
+    "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.3.0",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
+    "ngrx-store-freeze": "^0.1.9",
     "protractor": "~5.1.2",
-    "ts-node": "^3.0.6",
-    "tslint": "^5.4.3",
-    "typescript": "^2.3.4"
+    "ts-node": "^3.3.0",
+    "tslint": "^5.5.0",
+    "typescript": "^2.4.2"
   }
 }
