@@ -19,7 +19,7 @@ if (environment.hmr) {
     hmrBootstrap(module, bootstrap);
   } else {
     console.error('HMR is not enabled for webpack-dev-server!');
-    console.info('Are you using the --hmr flag for ng serve?');
+    console.log('Are you using the --hmr flag for ng serve?');
   }
 } else {
   bootstrap();
